requirements.txt
requests
xmltodict
bs4
python-dotenv